check for winner, update wins, add better reset
The game will now check for a winner each time a player clicks a square. When a player wins their win count is updated. The reset button will not reload the page when it is clicked when the board it empty to fully reset the game. The change names functionality will always update the turn message to be accurate to the current player names.